Extract Assessor: side-effect-free Phase 2/3 assessment
Dissolve the pipeline's Phase 2/3 responsibility into a dedicated Assessor
collaborator (interface + DefaultAssessor). assess(ctx, snapshot) -> Assessment
runs labeling (Phase 2) plus risk/rules/evidence (Phase 3) against a state
snapshot with no mutation or persistence, returning the SessionMeta verdict and
any deferred MCP write prescriptions.

Both the shield's preflight simulation and the monitor's post-commit path now
share exactly one assessment implementation instead of duplicating the
label -> phase3 -> build-meta sequence. Moves _phase3, _render_transform,
_build_escalation, _sanitize_args and _with_snapshot out of the god class;
_infer_phase (Phase 1) stays for the monitor.

Full suite green (1884 passed, 4 skipped); ruff clean.
